backend/requirements.txt
# FastAPI Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0
mysql-connector-python>=8.3.0
pymysql>=1.1.0
cryptography>=42.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.1.2

# HTTP & API
httpx>=0.26.0
requests>=2.31.0
aiohttp>=3.9.0

# Environment & Configuration
python-dotenv>=1.0.0

# Data Validation & Serialization
email-validator>=2.1.0
orjson>=3.9.0

# CORS & Middleware
python-cors>=1.0.0

# Logging & Monitoring
loguru>=0.7.2
python-json-logger>=2.0.7

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Development Tools
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0

# Google Books API Integration
google-api-python-client>=2.115.0

# Data Processing (from existing scripts)
pandas>=2.1.0
numpy>=1.24.0

# Async Support
asyncio>=3.4.3

# CLI Tools
click>=8.1.0
rich>=13.5.0

